% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionWelch.R
\name{welch}
\alias{welch}
\title{welch}
\usage{
welch(x, y)
}
\arguments{
\item{x, y}{muestras en forma de vectores numericos cuyos tamanios han de ser
menores que 30}
}
\description{
La funcion welch recoge dos muestras de tamanios menores que 30 y devuelve un
delta, el cual se utiliza para determinar los grados de libertad adecuados
para una t de student a la hora de realizar la aproximacion de Welch. Ahora,
no te digo yo que no sino es que si.
}
\examples{
#Testeo con las muestras presentes en los apuntes del tema 3 sobre intervalos de confianza.
x <- c(31, 18, 17, 16, 37, 16, 32, 13, 14, 49, 25, 19, 13, 32, 27)
y <- c(15, 17, 13, 25, 22, 20, 24, 12, 23, 15, 20, 18)

welch(x, y)
}
